a test hit a relationship unification failure - handle that case
