+++ crates/ndc-query-plan/src/plan_for_query_request/helpers.rs
@@ -1,6 +1,7 @@
 use std::collections::BTreeMap;
 
 use ndc_models as ndc;
+
 use crate as plan;
 
 use super::query_plan_error::QueryPlanError;
